// web/spectron/firebase-auth/lib/firebase.d.ts
declare namespace firebase {
    type NextFn<T> = (value: T) => void;
    type ErrorFn<E = Error> = (error: E) => void;
    type CompleteFn = () => void;

    interface FirebaseError {
        code: string;
        message: string;
        name: string;
        stack?: string;
    }

    interface Observer<T, E = Error> {
        next: NextFn<T>;
        error: ErrorFn<E>;
        complete: CompleteFn;
    }

    var SDK_VERSION: string;

    type Unsubscribe = () => void;

    interface User extends firebase.UserInfo {
        delete(): Promise<void>;
        emailVerified: boolean;
        getIdTokenResult(
            forceRefresh?: boolean
        ): Promise<firebase.auth.IdTokenResult>;
        getIdToken(forceRefresh?: boolean): Promise<string>;
        isAnonymous: boolean;
        linkAndRetrieveDataWithCredential(
            credential: firebase.auth.AuthCredential
        ): Promise<firebase.auth.UserCredential>;
        linkWithCredential(
            credential: firebase.auth.AuthCredential
        ): Promise<firebase.User>;
        linkWithPhoneNumber(
            phoneNumber: string,
            applicationVerifier: firebase.auth.ApplicationVerifier
        ): Promise<firebase.auth.ConfirmationResult>;
        linkWithPopup(
            provider: firebase.auth.AuthProvider
        ): Promise<firebase.auth.UserCredential>;
        linkWithRedirect(provider: firebase.auth.AuthProvider): Promise<void>;
        metadata: firebase.auth.UserMetadata;
        phoneNumber: string | null;
        providerData: (firebase.UserInfo | null)[];
        reauthenticateAndRetrieveDataWithCredential(
            credential: firebase.auth.AuthCredential
        ): Promise<firebase.auth.UserCredential>;
        reauthenticateWithCredential(
            credential: firebase.auth.AuthCredential
        ): Promise<void>;
        reauthenticateWithPhoneNumber(
            phoneNumber: string,
            applicationVerifier: firebase.auth.ApplicationVerifier
        ): Promise<firebase.auth.ConfirmationResult>;
        reauthenticateWithPopup(
            provider: firebase.auth.AuthProvider
        ): Promise<firebase.auth.UserCredential>;
        reauthenticateWithRedirect(
            provider: firebase.auth.AuthProvider
        ): Promise<void>;
        refreshToken: string;
        reload(): Promise<void>;
        sendEmailVerification(
            actionCodeSettings?: firebase.auth.ActionCodeSettings | null
        ): Promise<void>;
        toJSON(): Object;
        unlink(providerId: string): Promise<firebase.User>;
        updateEmail(newEmail: string): Promise<void>;
        updatePassword(newPassword: string): Promise<void>;
        updatePhoneNumber(
            phoneCredential: firebase.auth.AuthCredential
        ): Promise<void>;
        updateProfile(profile: {
            displayName: string | null;
            photoURL: string | null;
        }): Promise<void>;
    }

    interface UserInfo {
        displayName: string | null;
        email: string | null;
        phoneNumber: string | null;
        photoURL: string | null;
        providerId: string;
        uid: string;
    }

    function app(name?: string): firebase.app.App;

    var apps: firebase.app.App[];

    function auth(app?: firebase.app.App): firebase.auth.Auth;

    function database(app?: firebase.app.App): firebase.database.Database;

    function initializeApp(options: Object, name?: string): firebase.app.App;

    function messaging(app?: firebase.app.App): firebase.messaging.Messaging;

    function storage(app?: firebase.app.App): firebase.storage.Storage;

    function firestore(app?: firebase.app.App): firebase.firestore.Firestore;

    function functions(app?: firebase.app.App): firebase.functions.Functions;
}

declare namespace firebase.app {
    interface App {
        auth(): firebase.auth.Auth;
        database(url?: string): firebase.database.Database;
        delete(): Promise<any>;
        messaging(): firebase.messaging.Messaging;
        name: string;
        options: Object;
        storage(url?: string): firebase.storage.Storage;
        firestore(): firebase.firestore.Firestore;
        functions(region?: string): firebase.functions.Functions;
    }
}
// ...
    export class Functions {
        private constructor();
        httpsCallable(name: string): HttpsCallable;
    }
// ...
    type ActionCodeSettings = {
        android?: {
            installApp?: boolean;
            minimumVersion?: string;
            packageName: string;
        };
        handleCodeInApp?: boolean;
        iOS?: { bundleId: string };
        url: string;
    };

    type AdditionalUserInfo = {
        isNewUser: boolean;
        profile: Object | null;
        providerId: string;
        username?: string | null;
    };

    interface ApplicationVerifier {
        type: string;
        verify(): Promise<string>;
    }

    interface AuthSettings {
        appVerificationDisabledForTesting: boolean;
    }

    interface Auth {
        app: firebase.app.App;
        applyActionCode(code: string): Promise<void>;
        checkActionCode(code: string): Promise<firebase.auth.ActionCodeInfo>;
        confirmPasswordReset(code: string, newPassword: string): Promise<void>;
        createUserAndRetrieveDataWithEmailAndPassword(
            email: string,
            password: string
        ): Promise<firebase.auth.UserCredential>;
        createUserWithEmailAndPassword(
            email: string,
            password: string
        ): Promise<firebase.auth.UserCredential>;
        currentUser: firebase.User | null;
        fetchProvidersForEmail(email: string): Promise<Array<string>>;
        fetchSignInMethodsForEmail(email: string): Promise<Array<string>>;
        isSignInWithEmailLink(emailLink: string): boolean;
        getRedirectResult(): Promise<firebase.auth.UserCredential>;
        languageCode: string | null;
        settings: firebase.auth.AuthSettings;
        onAuthStateChanged(
            nextOrObserver:
                | firebase.Observer<any>
                | ((a: firebase.User | null) => any),
            error?: (a: firebase.auth.Error) => any,
            completed?: firebase.Unsubscribe
        ): firebase.Unsubscribe;
        onIdTokenChanged(
            nextOrObserver:
                | firebase.Observer<any>
                | ((a: firebase.User | null) => any),
            error?: (a: firebase.auth.Error) => any,
            completed?: firebase.Unsubscribe
        ): firebase.Unsubscribe;
        sendSignInLinkToEmail(
            email: string,
            actionCodeSettings: firebase.auth.ActionCodeSettings
        ): Promise<void>;
        sendPasswordResetEmail(
            email: string,
            actionCodeSettings?: firebase.auth.ActionCodeSettings | null
        ): Promise<void>;
        setPersistence(persistence: firebase.auth.Auth.Persistence): Promise<void>;
        signInAndRetrieveDataWithCredential(
            credential: firebase.auth.AuthCredential
        ): Promise<firebase.auth.UserCredential>;
        signInAnonymously(): Promise<firebase.auth.UserCredential>;
        signInAnonymouslyAndRetrieveData(): Promise<firebase.auth.UserCredential>;
        signInWithCredential(
            credential: firebase.auth.AuthCredential
        ): Promise<firebase.User>;
        signInWithCustomToken(token: string): Promise<firebase.auth.UserCredential>;
        signInAndRetrieveDataWithCustomToken(
            token: string
        ): Promise<firebase.auth.UserCredential>;
        signInWithEmailAndPassword(
            email: string,
            password: string
        ): Promise<firebase.auth.UserCredential>;
        signInAndRetrieveDataWithEmailAndPassword(
            email: string,
            password: string
        ): Promise<firebase.auth.UserCredential>;
        signInWithPhoneNumber(
            phoneNumber: string,
            applicationVerifier: firebase.auth.ApplicationVerifier
        ): Promise<firebase.auth.ConfirmationResult>;
        signInWithEmailLink(
            email: string,
            emailLink?: string
        ): Promise<firebase.auth.UserCredential>;
        signInWithPopup(
            provider: firebase.auth.AuthProvider
        ): Promise<firebase.auth.UserCredential>;
        signInWithRedirect(provider: firebase.auth.AuthProvider): Promise<void>;
        signOut(): Promise<void>;
        updateCurrentUser(user: firebase.User | null): Promise<void>;
        useDeviceLanguage(): void;
        verifyPasswordResetCode(code: string): Promise<string>;
    }

    interface AuthCredential {
        providerId: string;
        signInMethod: string;
    }

    interface AuthProvider {
        providerId: string;
    }

    interface ConfirmationResult {
        confirm(verificationCode: string): Promise<firebase.auth.UserCredential>;
        verificationId: string;
    }
// ...
    interface Error {
        code: string;
        message: string;
    }
// ...
    interface IdTokenResult {
        token: string;
        expirationTime: string;
        authTime: string;
        issuedAtTime: string;
        signInProvider: string | null;
        claims: {
            [key: string]: any;
        };
    }
// ...
    type UserCredential = {
        additionalUserInfo?: firebase.auth.AdditionalUserInfo | null;
        credential: firebase.auth.AuthCredential | null;
        operationType?: string | null;
        user: firebase.User | null;
    };

    interface UserMetadata {
        creationTime?: string;
        lastSignInTime?: string;
    }
// ...
    interface Database {
        app: firebase.app.App;
        goOffline(): any;
        goOnline(): any;
        ref(path?: string): firebase.database.Reference;
        refFromURL(url: string): firebase.database.Reference;
    }